… execution/binary/index.go
… execution/binary/vector.go
… go.sum