* rewritten binary operators ( really no good reason; it just kinda happened )
* added support for "and, or, unless" ( prepared support for native histograms but that will be followup )
* fixed some bugs in our duplicate label detection logic
* fixed minor bug  in error reporting for hash collision on left side in a one-to-one matching