@MichaHoffmann 2023-09-30
```
$ benchstat main.bench pr.bench
goos: linux
goarch: amd64
pkg: github.com/thanos-io/promql-engine/engine
cpu: 11th Gen Intel(R) Core(TM) i5-1135G7 @ 2.40GHz
                                                                │  main.bench   │              pr.bench              │
                                                                │    sec/op     │    sec/op     vs base              │
RangeQuery/binary_operation_with_one_to_one/new_engine-8           53.48m ± ∞ ¹   48.23m ± ∞ ¹  -9.82% (p=0.008 n=5)
RangeQuery/binary_operation_with_many_to_one/new_engine-8          125.8m ± ∞ ¹   121.6m ± ∞ ¹  -3.36% (p=0.008 n=5)
RangeQuery/binary_operation_with_vector_and_scalar/new_engine-8   102.98m ± ∞ ¹   99.02m ± ∞ ¹  -3.84% (p=0.008 n=5)
geomean                                                            88.50m         83.43m        -5.72%
¹ need >= 6 samples for confidence interval at level 0.95

                                                                │  main.bench   │               pr.bench               │
                                                                │     B/op      │     B/op       vs base               │
RangeQuery/binary_operation_with_one_to_one/new_engine-8          16.72Mi ± ∞ ¹   12.62Mi ± ∞ ¹  -24.54% (p=0.008 n=5)
RangeQuery/binary_operation_with_many_to_one/new_engine-8         42.33Mi ± ∞ ¹   31.58Mi ± ∞ ¹  -25.39% (p=0.008 n=5)
RangeQuery/binary_operation_with_vector_and_scalar/new_engine-8   28.23Mi ± ∞ ¹   28.15Mi ± ∞ ¹        ~ (p=0.310 n=5)
geomean                                                           27.14Mi         22.39Mi        -17.51%
¹ need >= 6 samples for confidence interval at level 0.95

                                                                │  main.bench  │              pr.bench               │
                                                                │  allocs/op   │  allocs/op    vs base               │
RangeQuery/binary_operation_with_one_to_one/new_engine-8          49.58k ± ∞ ¹   37.30k ± ∞ ¹  -24.77% (p=0.008 n=5)
RangeQuery/binary_operation_with_many_to_one/new_engine-8         91.59k ± ∞ ¹   69.48k ± ∞ ¹  -24.13% (p=0.008 n=5)
RangeQuery/binary_operation_with_vector_and_scalar/new_engine-8   55.46k ± ∞ ¹   55.43k ± ∞ ¹        ~ (p=0.310 n=5)
geomean                                                           63.15k         52.37k        -17.07%
¹ need >= 6 samples for confidence interval at level 0.95
```